requirements.txt
Brotli==1.1.0
MarkupSafe==3.0.2
PySocks==1.7.1
PyYAML==6.0.2
archspec==0.2.5
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
boltons==24.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.2
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
evalidate==2.0.5
exceptiongroup==1.3.0
executing==2.2.0
expecttest==0.3.0
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.5.1
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
hypothesis==6.135.0
idna==3.10
importlib_resources==6.5.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
libarchive-c==5.3
libmambapy==2.1.1
lief==0.16.4
lintrunner==0.12.7
matplotlib-inline==0.1.7
menuinst==2.2.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
optree==0.16.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.1.1
pkginfo==1.12.1.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat==0.6.6
pycparser==2.22
Pygments==2.19.1
python-etcd==0.4.5
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.12
setuptools==75.8.2
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.7
stack_data==0.6.3
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchelastic==0.2.2
torchvision==0.22.1+cu118
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
truststore==0.10.1
types-dataclasses==0.6.6
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.22.0
zstandard==0.23.0
antlr4-python3-runtime==4.9.3
omegaconf==2.3.0
hydra-core==1.3.2
typing_extensions==4.15.0
smmap==5.0.2
sentry-sdk==2.49.0
annotated-types==0.7.0
typing-inspection==0.4.2
pydantic_core==2.41.5
gitdb==4.0.12
pydantic==2.12.5
GitPython==3.1.46
wandb==0.24.0
absl-py==2.3.1
grpcio==1.76.0
Markdown==3.10
protobuf==6.33.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
Werkzeug==3.1.3
bvhio==1.5.4
debugpy==1.8.17
pyglm==2.8.2
spatial-transform==1.3.3
body_visualizer==1.1.0
chumpy==0.70
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
freetype-py==2.5.1
imageio==2.37.0
kiwisolver==1.4.9
matplotlib==3.10.6
PyOpenGL==3.1.0
pyparsing==3.2.5
pyrender==0.1.45
python-dateutil==2.9.0.post0
smplx==0.1.28
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
dotmap==1.3.30
frozenlist==1.7.0
human_body_prior==2.2.2.0
lightning-utilities==0.15.2
multidict==6.6.4
opencv-python==4.12.0.88
propcache==0.4.0
psbody-mesh==0.4
pyglet==1.5.31
pytorch-lightning==2.5.5
pyzmq==27.1.0
scipy==1.16.2
torchmetrics==1.8.2
transforms3d==0.4.2
trimesh==4.8.3
yarl==1.20.1
